[VPNLINUX-820] Fix WG connection glitch
The route added to be able to connect to the VPN server with the
KS enabled is not available straight away. Due to this the
VPN connection was first failing (because the route was not
available) and then it succeeded on the next try, when the
reconnector quicked in (because to route was available then).

To know when the route is effective, I tried to monitor the
device interface and its active connection for state changes
but none were triggered.

Instead, I implemented a polling mechanism that runs the `ip route`
command and waits for the route to appear in the Linux routing
table before starting the VPN connection.